# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Japonica CRM",
    "description": "Japonica CRM: modular FastAPI backend for auth, orders, warehouse, franchisee onboarding, notifications and analytics.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/notifications/telegram/link": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Create Telegram Link",
        "operationId": "create_telegram_link_api_v1_notifications_telegram_link_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLinkResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/preferences": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Notification Preferences",
        "operationId": "notification_preferences_api_v1_notifications_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "notifications"
        ],
        "summary": "Patch Notification Preferences",
        "operationId": "patch_notification_preferences_api_v1_notifications_preferences_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/telegram/status": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Telegram Status",
        "operationId": "telegram_status_api_v1_notifications_telegram_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/telegram/unlink": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Telegram Unlink",
        "operationId": "telegram_unlink_api_v1_notifications_telegram_unlink_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/telegram/webhook": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Telegram Webhook",
        "operationId": "telegram_webhook_api_v1_notifications_telegram_webhook_post",
        "parameters": [
          {
            "name": "X-Telegram-Bot-Api-Secret-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Bot-Api-Secret-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramWebhookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationPreferences": {
        "properties": {
          "order_created": {
            "type": "boolean",
            "title": "Order Created",
            "default": true
          },
          "order_cancelled": {
            "type": "boolean",
            "title": "Order Cancelled",
            "default": true
          },
          "low_stock": {
            "type": "boolean",
            "title": "Low Stock",
            "default": true
          },
          "franchisee_stage_changed": {
            "type": "boolean",
            "title": "Franchisee Stage Changed",
            "default": true
          },
          "franchisee_task_changed": {
            "type": "boolean",
            "title": "Franchisee Task Changed",
            "default": true
          },
          "franchisee_task_overdue": {
            "type": "boolean",
            "title": "Franchisee Task Overdue",
            "default": true
          },
          "weekly_revenue_report": {
            "type": "boolean",
            "title": "Weekly Revenue Report",
            "default": true
          }
        },
        "type": "object",
        "title": "NotificationPreferences"
      },
      "NotificationPreferencesUpdate": {
        "properties": {
          "order_created": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Created"
          },
          "order_cancelled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Cancelled"
          },
          "low_stock": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Stock"
          },
          "franchisee_stage_changed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Franchisee Stage Changed"
          },
          "franchisee_task_changed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Franchisee Task Changed"
          },
          "franchisee_task_overdue": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Franchisee Task Overdue"
          },
          "weekly_revenue_report": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Revenue Report"
          }
        },
        "type": "object",
        "title": "NotificationPreferencesUpdate"
      },
...
      "OperationStatusResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "OperationStatusResponse"
      },
...
      "TelegramLinkResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "instructions": {
            "type": "string",
            "title": "Instructions"
          }
        },
        "type": "object",
        "required": [
          "code",
          "expires_in",
          "instructions"
        ],
        "title": "TelegramLinkResponse"
      },
      "TelegramStatusResponse": {
        "properties": {
          "linked": {
            "type": "boolean",
            "title": "Linked"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          }
        },
        "type": "object",
        "required": [
          "linked",
          "chat_id"
        ],
        "title": "TelegramStatusResponse"
      },
      "TelegramWebhookCallbackQuery": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TelegramWebhookMessage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "TelegramWebhookCallbackQuery"
      },
      "TelegramWebhookChat": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "TelegramWebhookChat"
      },
      "TelegramWebhookMessage": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "chat": {
            "$ref": "#/components/schemas/TelegramWebhookChat"
          }
        },
        "type": "object",
        "required": [
          "chat"
        ],
        "title": "TelegramWebhookMessage"
      },
      "TelegramWebhookUpdate": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TelegramWebhookMessage"
              },
              {
                "type": "null"
              }
            ]
          },
          "callback_query": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TelegramWebhookCallbackQuery"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "TelegramWebhookUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}